move items code from prototype
